+++ app/src/main/java/com/barryburgle/gameapp/ui/input/dialog/InputDialogConstant.kt
@@ -1,0 +1,7 @@
+package com.barryburgle.gameapp.ui.input.dialog
+
+class InputDialogConstant {
+    companion object {
+        const val EMPTY_STICKING_POINTS = "-"
+    }
+}